+++ src/main/java/dk/kb/datahandler/webservice/ContextListener.java
@@ -14,6 +14,7 @@ import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
 
 import dk.kb.datahandler.config.ServiceConfig;
+import dk.kb.datahandler.preservica.client.DsPreservicaClient;
 import dk.kb.util.BuildInfoManager;
 import dk.kb.util.Files;
 import dk.kb.util.Resolver;
@@ -67,6 +68,11 @@ public class ContextListener implements ServletContextListener {
             String configFile = (String) ctx.lookup("java:/comp/env/application-config");
             //TODO this should not refer to something in template. Should we perhaps use reflection here?
             ServiceConfig.initialize(configFile);
+            // Early initialization of preservica client.
+            DsPreservicaClient.init(ServiceConfig.getPreservicaUrl(), ServiceConfig.getPreservicaUser(),
+                    ServiceConfig.getPreservicaPassword(), ServiceConfig.getPreservicaKeepAliveSeconds());
+            log.info("Initialized Preservica Client to the following URL: '{}' for the user: '{}'",
+                    ServiceConfig.getPreservicaUrl(), ServiceConfig.getPreservicaUser());
         } catch (NamingException e) {
             throw new RuntimeException("Failed to lookup settings", e);
         } catch (IOException e) {