+++ src/test/java/dk/kb/datahandler/preservica/PreservicaDataTest.java
@@ -1,4 +1,4 @@
-package dk.kb.datahandler.pvica;
+package dk.kb.datahandler.preservica;
 
 import dk.kb.datahandler.model.v1.OaiTargetDto;
 import dk.kb.datahandler.oai.*;
@@ -22,7 +22,7 @@ import java.util.ArrayList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class PvicaDataTest {
+public class PreservicaDataTest {
     
     @Test
     public void testDeliverableUnitNameSpaceFix() throws Exception{
@@ -94,29 +94,7 @@ public class PvicaDataTest {
     public void testRemovalOfCollectionsAndPreservationManifestations() throws ApiException {
         // We don't need a working storage to test this functionality. Therefore this mock
         DsStorageClient mockedStorage = Mockito.mock(DsStorageClient.class);
-        // Create OAI-PMH test records.
-        OaiRecord deliverableUnit1 = new OaiRecord();
-        deliverableUnit1.setId("oai:du:12345678-test-test-test-testtest1111");
-        deliverableUnit1.setMetadata("<Metadata schemaURI=\"http://www.pbcore.org/PBCore/PBCoreNamespace.html\">"+
-                                        "<formatMediaType>Sound</formatMediaType>");
-        OaiRecord deliverableUnit2 = new OaiRecord();
-        deliverableUnit2.setId("oai:du:12345678-test-test-test-testtest2222");
-        deliverableUnit2.setMetadata("<Metadata schemaURI=\"http://www.pbcore.org/PBCore/PBCoreNamespace.html\">"+
-                                        "<formatMediaType>Sound</formatMediaType>");
-        OaiRecord preservationManifestation = new OaiRecord();
-        preservationManifestation.setId("oai:man:12345678-test-test-test-testtest2222");
-        preservationManifestation.setMetadata("<ManifestationRelRef>1</ManifestationRelRef>");
-        OaiRecord collection = new OaiRecord();
-        collection.setId("oai:col:123456-test-1234");
-        // Create ArrayList of OAI records.
-        ArrayList<OaiRecord> oaiRecords = new ArrayList<>();
-        oaiRecords.add(collection);
-        oaiRecords.add(deliverableUnit1);
-        oaiRecords.add(deliverableUnit2);
-        oaiRecords.add(preservationManifestation);
-        // Create test OaiResponse.
-        OaiResponse testResponse = new OaiResponse();
-        testResponse.setRecords(oaiRecords);
+        OaiResponse testResponse = createPreserviceFiveOaiResponse();
 
         OaiResponseFilter oaiFilter = new OaiResponseFilterPreservicaFive("preservica", mockedStorage);
 
@@ -141,6 +119,7 @@ public class PvicaDataTest {
 
         assertEquals(RecordTypeDto.COLLECTION, resolvedType);
     }
+
     @Test
     public void testRecordTypeDU() {
         DsRecordDto collectionRecord = new DsRecordDto();
@@ -159,7 +138,6 @@ public class PvicaDataTest {
 
         assertEquals(RecordTypeDto.MANIFESTATION, resolvedType);
     }
-
     @Test
     public void testRecordForTransformationErrors() throws IOException, SAXException, ParserConfigurationException {
         // Read XML as Document
@@ -188,6 +166,58 @@ public class PvicaDataTest {
         //Tests
         assertEquals(RecordTypeDto.DELIVERABLEUNIT, resolvedType);
         assertEquals("ds.test", dsRecord.getOrigin());
+    }
+
+    /**
+     * Create an OAI test response containing two InformationObjects.
+     * @return OAI Response with two fake IO records.
+     */
+    private static OaiResponse createTestOaiResponseWithIOs() {
+        // Create OAI-PMH test records.
+        OaiRecord informationObject1 = new OaiRecord();
+        informationObject1.setId("oai:io:12345678-test-test-test-testtest1111");
+        informationObject1.setMetadata("<Metadata schemaUri=\"http://www.pbcore.org/PBCore/PBCoreNamespace.html\">"+
+                "<formatMediaType>Sound</formatMediaType>");
+        OaiRecord informationObject22 = new OaiRecord();
+        informationObject22.setId("oai:io:12345678-test-test-test-testtest2222");
+        informationObject22.setMetadata("<Metadata schemaUri=\"http://www.pbcore.org/PBCore/PBCoreNamespace.html\">"+
+                "<formatMediaType>Sound</formatMediaType>");
+
+        ArrayList<OaiRecord> oaiRecords = new ArrayList<>();
+        oaiRecords.add(informationObject1);
+        oaiRecords.add(informationObject22);
+        // Create test OaiResponse.
+        OaiResponse testResponse = new OaiResponse();
+        testResponse.setRecords(oaiRecords);
+        return testResponse;
+    }
+
+    private static OaiResponse createPreserviceFiveOaiResponse() {
+        // Create OAI-PMH test records.
+        OaiRecord deliverableUnit1 = new OaiRecord();
+
+        deliverableUnit1.setId("oai:du:12345678-test-test-test-testtest1111");
+        deliverableUnit1.setMetadata("<Metadata schemaURI=\"http://www.pbcore.org/PBCore/PBCoreNamespace.html\">"+
+                "<formatMediaType>Sound</formatMediaType>");
+        OaiRecord deliverableUnit2 = new OaiRecord();
+        deliverableUnit2.setId("oai:du:12345678-test-test-test-testtest2222");
+        deliverableUnit2.setMetadata("<Metadata schemaURI=\"http://www.pbcore.org/PBCore/PBCoreNamespace.html\">"+
+                "<formatMediaType>Sound</formatMediaType>");
+        OaiRecord preservationManifestation = new OaiRecord();
+        preservationManifestation.setId("oai:man:12345678-test-test-test-testtest2222");
+        preservationManifestation.setMetadata("<ManifestationRelRef>1</ManifestationRelRef>");
+        OaiRecord collection = new OaiRecord();
+        collection.setId("oai:col:123456-test-1234");
+        // Create ArrayList of OAI records.
+        ArrayList<OaiRecord> oaiRecords = new ArrayList<>();
+        oaiRecords.add(collection);
+        oaiRecords.add(deliverableUnit1);
+        oaiRecords.add(deliverableUnit2);
+        oaiRecords.add(preservationManifestation);
+        // Create test OaiResponse.
+        OaiResponse testResponse = new OaiResponse();
+        testResponse.setRecords(oaiRecords);
+        return testResponse;
     }
 
 }